requirements.txt
line-bot-sdk==3.16.3
Flask==3.1.0